+++ edison-togglz/src/main/java/de/otto/edison/togglz/configuration/MongoTogglzConfiguration.java
@@ -1,10 +1,12 @@
-package de.otto.edison.mongo.configuration;
+package de.otto.edison.togglz.configuration;
 
 import com.mongodb.client.MongoDatabase;
-import de.otto.edison.mongo.togglz.MongoTogglzRepository;
+import de.otto.edison.mongo.configuration.MongoProperties;
 import de.otto.edison.togglz.FeatureClassProvider;
+import de.otto.edison.togglz.mongo.MongoTogglzRepository;
 import org.slf4j.Logger;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -14,15 +16,18 @@ import org.togglz.core.user.UserProvider;
 import static org.slf4j.LoggerFactory.getLogger;
 
 @Configuration
-@ConditionalOnClass(name = "de.otto.edison.togglz.configuration.TogglzConfiguration")
 @EnableConfigurationProperties(MongoProperties.class)
+@ConditionalOnProperty(prefix = "edison.togglz", name = "mongo.enabled", havingValue = "true")
+@ConditionalOnBean(type = "com.mongodb.MongoClient")
 public class MongoTogglzConfiguration {
 
     private static final Logger LOG = getLogger(MongoTogglzConfiguration.class);
 
     @Bean
-    StateRepository stateRepository(final MongoDatabase mongoDatabase, final FeatureClassProvider featureClassProvider,
-                                    final UserProvider userProvider, final MongoProperties mongoProperties) {
+    StateRepository stateRepository(final MongoDatabase mongoDatabase,
+                                    final FeatureClassProvider featureClassProvider,
+                                    final UserProvider userProvider,
+                                    final MongoProperties mongoProperties) {
         LOG.info("===============================");
         LOG.info("Using MongoTogglzRepository with " + mongoDatabase.getClass().getSimpleName() + " MongoDatabase impl.");
         LOG.info("===============================");