+++ edison-togglz/src/main/java/de/otto/edison/togglz/mongo/MongoTogglzRepository.java
@@ -1,4 +1,4 @@
-package de.otto.edison.mongo.togglz;
+package de.otto.edison.togglz.mongo;
 
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;