+++ edison-togglz/src/test/java/de/otto/edison/togglz/DefaultTogglzConfigTest.java
@@ -1,39 +1,38 @@
 package de.otto.edison.togglz;
 
-import de.otto.edison.testsupport.applicationdriver.SpringTestBase;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.Profile;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.togglz.core.manager.TogglzConfig;
 import org.togglz.core.repository.cache.CachingStateRepository;
 
 import static de.otto.edison.testsupport.dsl.Then.assertThat;
-import static org.hamcrest.Matchers.instanceOf;
-import static org.hamcrest.Matchers.not;
-import static org.hamcrest.Matchers.nullValue;
-import static org.hamcrest.Matchers.typeCompatibleWith;
+import static org.hamcrest.Matchers.*;
 import static org.hamcrest.core.Is.is;
+import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
 
-public class DefaultTogglzConfigTest extends SpringTestBase {
+@ExtendWith(SpringExtension.class)
+@SpringBootTest(webEnvironment = RANDOM_PORT, classes = {TestServer.class})
+@Profile("test")
+public class DefaultTogglzConfigTest {
 
+    @Autowired
     private TogglzConfig togglzConfig;
-
-    @BeforeEach
-    public void setUp() {
-        togglzConfig = applicationContext().getBean(TogglzConfig.class);
-
-    }
 
     @Test
     public void shouldCreateTogglzConfigBySpring() {
         assertThat(togglzConfig, is(not(nullValue())));
         assertThat(togglzConfig.getFeatureClass(), typeCompatibleWith(TestFeatures.class));
-        assertThat(togglzConfig.getStateRepository(),is(not(nullValue())));
-        assertThat(togglzConfig.getStateRepository(),is(instanceOf(CachingStateRepository.class)));
-        assertThat(togglzConfig.getUserProvider(),is(not(nullValue())));
+        assertThat(togglzConfig.getStateRepository(), is(not(nullValue())));
+        assertThat(togglzConfig.getStateRepository(), is(instanceOf(CachingStateRepository.class)));
+        assertThat(togglzConfig.getUserProvider(), is(not(nullValue())));
     }
 
     @Test
     public void shouldProvideToggleStateWhichIsActiveByDefaultInTests() {
-        assertThat(TestFeatures.TEST_FEATURE.isActive(),is(true));
+        assertThat(TestFeatures.TEST_FEATURE.isActive(), is(true));
     }
 }